session config extracted
