+++ Refactoring/app.js
@@ -2,22 +2,16 @@ const path = require('path');
 
 const express = require('express');
 const session = require('express-session');
-const mongodbStore = require('connect-mongodb-session');
 const csrf = require('csurf');
 
+const sessionConfig = require('./config/session');
 const db = require('./data/database');
 const authRoutes = require('./routes/auth');
 const blogRoutes = require('./routes/blog');
 
-const MongoDBStore = mongodbStore(session);
+const mongoDbSessionStore = sessionConfig.createSessionStore(session);
 
 const app = express();
-
-const sessionStore = new MongoDBStore({
-  uri: 'mongodb://127.0.0.1:27017',
-  databaseName: 'auth-demo',
-  collection: 'sessions'
-});
 
 app.set('view engine', 'ejs');
 app.set('views', path.join(__dirname, 'views'));
@@ -25,15 +19,7 @@ app.set('views', path.join(__dirname, 'views'));
 app.use(express.static('public'));
 app.use(express.urlencoded({ extended: false }));
 
-app.use(session({
-  secret: 'super-secret',
-  resave: false,
-  saveUninitialized: false,
-  store: sessionStore,
-  cookie: {
-    maxAge: 2 * 24 * 60 * 60 * 1000
-  }
-}));
+app.use(session(sessionConfig.createSessionConfig(mongoDbSessionStore)));
 app.use(csrf());
 
 app.use(async function(req, res, next) {
